requirements.txt
python-dotenv
selenium
beautifulsoup4
python-telegram-bot
Flask
